sharemyshows/backend/requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7

# Redis
redis==5.0.1
Flask-Session==0.5.0

# Authentication
Flask-JWT-Extended==4.5.3
bcrypt==4.1.1

# Validation
marshmallow==3.20.1
email-validator==2.1.0

# Environment
python-dotenv==1.0.0

# WSGI Server (for production)
gunicorn==21.2.0

# Background Tasks (Phase 2)
# celery==5.3.4

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
faker==20.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.0
pylint==3.0.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
